+++ src/interfaces/I_PokemonItem.ts
@@ -1,0 +1,4 @@
+export default interface I_PokemonItem {
+  name: string;
+  url: string;
+}
